Change Booking to Transaction Controller, and fix upload photo for edit user bug
