+++ routes/index.js
@@ -12,6 +12,7 @@ const Passenger = require("./passengerRouter");
 const Ticket = require("./ticketRouter");
 const Booking = require("./bookingRouter");
 const Seat = require("./seatRoutes");
+const Transaction = require("./transactionRouter");
 router.use("/api-docs", swaggerUI.serve);
 router.use("/api-docs", swaggerUI.setup(swaggerDocument));
 
@@ -22,7 +23,7 @@ router.use("/api/v1/airline", Airline);
 router.use("/api/v1/user", User);
 router.use("/api/v1/passenger", Passenger);
 router.use("/api/v1/ticket", Ticket);
-router.use("/api/v1/payment", Booking);
+router.use("/api/v1/payment", Transaction);
 router.use("/api/v1/seat", Seat);
 
 module.exports = router;
